refactor: extract asset-detail sub-components into page folder
Closes #363
